requirements.txt
paho-mqtt==1.2
requests==2.10.0
